Eliminate compiler warnings for MSVC
Eliminate compiler warnings for MSVC, mostly by adding static_cast.
Fixes #65.

+++ re2/nfa.cc
@@ -608,7 +608,7 @@ bool NFA::Search(const StringPiece& text, const StringPiece& const_context,
 
   if (matched_) {
     for (int i = 0; i < nsubmatch; i++)
-      submatch[i].set(match_[2*i], match_[2*i+1] - match_[2*i]);
+       submatch[i].set(match_[2 * i], static_cast<int>(match_[2 * i + 1] - match_[2 * i]));
     if (Debug)
       fprintf(stderr, "match (%d,%d)\n",
               static_cast<int>(match_[0] - btext_),
