+++ re2/testing/tester.cc
@@ -393,8 +393,8 @@ void TestInstance::RunSearch(Engine type,
         re_anchor = RE2::ANCHOR_BOTH;
 
       result->matched = re2_->Match(context,
-                                    text.begin() - context.begin(),
-                                    text.end() - context.begin(),
+                                    static_cast<int>(text.begin() - context.begin()),
+                                    static_cast<int>(text.end() - context.begin()),
                                     re_anchor, result->submatch, nsubmatch);
       result->have_submatch = nsubmatch > 0;
       break;
